third_party/WebKit/LayoutTests/inspector-protocol/debugger/domdebugger-script-blocked-by-csp-expected.txt

CONSOLE ERROR: line 2: Refused to execute inline event handler because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 4: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-bhHHL3z2vDgxUt0W3dWQOrprscmda2Y5pLsLg4GF+pI='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 5: Refused to execute JavaScript URL because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

Button

-------
blockedEventHandler
Paused at: blockedEventHandler@1
Reason: EventListener; Data:
{
    directiveText : script-src 'self'
    eventName : instrumentation:scriptBlockedByCSP
}

-------
blockedScriptInjection
Paused at: blockedScriptInjection@3
Reason: EventListener; Data:
{
    directiveText : script-src 'self'
    eventName : instrumentation:scriptBlockedByCSP
}

-------
blockedScriptUrl
Paused at: blockedScriptUrl@4
Reason: EventListener; Data:
{
    directiveText : script-src 'self'
    eventName : instrumentation:scriptBlockedByCSP
}
